requirements.txt
bs4==0.0.1 
requests==2.23.0
python-telegram-bot==12.4.2